requirements.txt
environs==9.5.0
python-telegram-bot==13.12
redis==4.3.3
vk-api==11.9.8